Matrix: Replaced size_t with uint32_t

+++ include/sfz/math/Matrix.inl
@@ -21,40 +21,40 @@ namespace sfz {
 // Constructors & destructors
 // ------------------------------------------------------------------------------------------------
 
-template<typename T, size_t M, size_t N>
+template<typename T, uint32_t M, uint32_t N>
 Matrix<T,M,N>::Matrix(const T* arrayPtr, bool rowMajorData) noexcept
 {
 	static_assert(sizeof(elements) == (M*N*sizeof(T)), "Matrix is padded.");
 	T* data = this->data();
 	if (rowMajorData) {
-		for (size_t i = 0; i < M; ++i) {
-			for (size_t j = 0; j < N; ++j) {
+		for (uint32_t i = 0; i < M; ++i) {
+			for (uint32_t j = 0; j < N; ++j) {
 				data[j*M + i] = arrayPtr[i*N + j];
 			}
 		}
 	} else {
-		for (size_t i = 0; i < M*N; ++i) {
+		for (uint32_t i = 0; i < M*N; ++i) {
 			data[i] = arrayPtr[i];
 		}
 	}
 }
 
-template<typename T, size_t M, size_t N>
+template<typename T, uint32_t M, uint32_t N>
 Matrix<T,M,N>::Matrix(std::initializer_list<std::initializer_list<T>> list) noexcept
 {
 	static_assert(sizeof(elements) == (M*N*sizeof(T)), "Matrix is padded.");
 
-	for (size_t j = 0; j < N; ++j) {
-		for (size_t i = 0; i < M; ++i) {
+	for (uint32_t j = 0; j < N; ++j) {
+		for (uint32_t i = 0; i < M; ++i) {
 			elements[j][i] = T(0);
 		}
 	}
 
 	sfz_assert_debug(list.size() <= M);
 	const std::initializer_list<T>* rowsPtr = list.begin();
-	for (size_t i = 0; i < list.size(); ++i) {
+	for (uint32_t i = 0; i < list.size(); ++i) {
 		sfz_assert_debug(rowsPtr[i].size() <= N);
-		for (size_t j = 0; j < rowsPtr[i].size(); ++j) {
+		for (uint32_t j = 0; j < rowsPtr[i].size(); ++j) {
 			elements[j][i] = rowsPtr[i].begin()[j];
 		}
 	}
@@ -63,24 +63,24 @@ Matrix<T,M,N>::Matrix(std::initializer_list<std::initializer_list<T>> list) noex
 // Member functions
 // ------------------------------------------------------------------------------------------------
 
-template<typename T, size_t M, size_t N>
-T& Matrix<T,M,N>::at(size_t i, size_t j) noexcept
+template<typename T, uint32_t M, uint32_t N>
+T& Matrix<T,M,N>::at(uint32_t i, uint32_t j) noexcept
 {
 	sfz_assert_debug(i < M);
 	sfz_assert_debug(j < N);
 	return elements[j][i];
 }
 
-template<typename T, size_t M, size_t N>
-T Matrix<T,M,N>::at(size_t i, size_t j) const noexcept
+template<typename T, uint32_t M, uint32_t N>
+T Matrix<T,M,N>::at(uint32_t i, uint32_t j) const noexcept
 {
 	sfz_assert_debug(i < M);
 	sfz_assert_debug(j < N);
 	return elements[j][i];
 }
 
-template<typename T, size_t M, size_t N>
-Vector<T,N> Matrix<T,M,N>::rowAt(size_t i) const noexcept
+template<typename T, uint32_t M, uint32_t N>
+Vector<T,N> Matrix<T,M,N>::rowAt(uint32_t i) const noexcept
 {
 	sfz_assert_debug(i < M);
 	Vector<T,N> row;
@@ -90,8 +90,8 @@ Vector<T,N> Matrix<T,M,N>::rowAt(size_t i) const noexcept
 	return row;
 }
 
-template<typename T, size_t M, size_t N>
-Vector<T,M> Matrix<T,M,N>::columnAt(size_t j) const noexcept
+template<typename T, uint32_t M, uint32_t N>
+Vector<T,M> Matrix<T,M,N>::columnAt(uint32_t j) const noexcept
 {
 	sfz_assert_debug(j < N);
 	Vector<T,M> column;
@@ -101,16 +101,16 @@ Vector<T,M> Matrix<T,M,N>::columnAt(size_t j) const noexcept
 	return column;
 }
 
-template<typename T, size_t M, size_t N>
-void Matrix<T,M,N>::set(size_t i, size_t j, T value) noexcept
+template<typename T, uint32_t M, uint32_t N>
+void Matrix<T,M,N>::set(uint32_t i, uint32_t j, T value) noexcept
 {
 	sfz_assert_debug(i < M);
 	sfz_assert_debug(j < N);
 	elements[j][i] = value;
 }
 
-template<typename T, size_t M, size_t N>
-void Matrix<T,M,N>::setRow(size_t i, const Vector<T,N>& row) noexcept
+template<typename T, uint32_t M, uint32_t N>
+void Matrix<T,M,N>::setRow(uint32_t i, const Vector<T,N>& row) noexcept
 {
 	sfz_assert_debug(i < M);
 	for (uint32_t j = 0; j < N; j++) {
@@ -118,8 +118,8 @@ void Matrix<T,M,N>::setRow(size_t i, const Vector<T,N>& row) noexcept
 	}
 }
 
-template<typename T, size_t M, size_t N>
-void Matrix<T,M,N>::setColumn(size_t j, const Vector<T,M>& column) noexcept
+template<typename T, uint32_t M, uint32_t N>
+void Matrix<T,M,N>::setColumn(uint32_t j, const Vector<T,M>& column) noexcept
 {
 	sfz_assert_debug(j < N);
 	for (uint32_t i = 0; i < M; i++) {
@@ -130,7 +130,7 @@ void Matrix<T,M,N>::setColumn(size_t j, const Vector<T,M>& column) noexcept
 // Matrix constants
 // ------------------------------------------------------------------------------------------------
 
-template<typename T, size_t M, size_t N>
+template<typename T, uint32_t M, uint32_t N>
 Matrix<T,M,N> ZERO_MATRIX() noexcept
 {
 	static const Matrix<T,M,N> ZERO = []{ Matrix<T,M,N> temp; fill(temp, T(0)); return temp; }();
@@ -140,34 +140,34 @@ Matrix<T,M,N> ZERO_MATRIX() noexcept
 // Matrix functions
 // ------------------------------------------------------------------------------------------------
 
-template<typename T, size_t M, size_t N>
+template<typename T, uint32_t M, uint32_t N>
 void fill(Matrix<T, M, N>& matrix, T value)
 {
-	for (size_t i = 0; i < M; ++i) {
-		for (size_t j = 0; j < N; ++j) {
+	for (uint32_t i = 0; i < M; ++i) {
+		for (uint32_t j = 0; j < N; ++j) {
 			matrix.elements[j][i] = value;
 		}
 	}
 }
 
-template<typename T, size_t M, size_t N>
+template<typename T, uint32_t M, uint32_t N>
 Matrix<T,M,N> elemMult(const Matrix<T,M,N>& lhs, const Matrix<T,M,N>& rhs) noexcept
 {
 	Matrix<T,M,N> resMatrix;
-	for (size_t i = 0; i < M; i++) {
-		for (size_t j = 0; j < N; j++) {
+	for (uint32_t i = 0; i < M; i++) {
+		for (uint32_t j = 0; j < N; j++) {
 			resMatrix.elements[j][i] = lhs.elements[j][i] * rhs.elements[j][i];
 		}
 	}
 	return resMatrix;
 }
 
-template<typename T, size_t M, size_t N>
+template<typename T, uint32_t M, uint32_t N>
 Matrix<T,N,M> transpose(const Matrix<T,M,N>& matrix) noexcept
 {
 	Matrix<T,N,M> resMatrix;
-	for (size_t i = 0; i < N; i++) {
-		for (size_t j = 0; j < M; j++) {
+	for (uint32_t i = 0; i < N; i++) {
+		for (uint32_t j = 0; j < M; j++) {
 			resMatrix.elements[j][i] = matrix.elements[i][j];
 		}
 	}
@@ -177,40 +177,40 @@ Matrix<T,N,M> transpose(const Matrix<T,M,N>& matrix) noexcept
 // Operators (arithmetic & assignment)
 // ------------------------------------------------------------------------------------------------
 
-template<typename T, size_t M, size_t N>
+template<typename T, uint32_t M, uint32_t N>
 Matrix<T,M,N>& operator+= (Matrix<T,M,N>& lhs, const Matrix<T,M,N>& rhs) noexcept
 {
-	for (size_t i = 0; i < M; ++i) {
-		for (size_t j = 0; j < N; ++j) {
+	for (uint32_t i = 0; i < M; ++i) {
+		for (uint32_t j = 0; j < N; ++j) {
 			lhs.elements[j][i] += rhs.elements[j][i];
 		}
 	}
 	return lhs;
 }
 
-template<typename T, size_t M, size_t N>
+template<typename T, uint32_t M, uint32_t N>
 Matrix<T,M,N>& operator-= (Matrix<T,M,N>& lhs, const Matrix<T,M,N>& rhs) noexcept
 {
-	for (size_t i = 0; i < M; ++i) {
-		for (size_t j = 0; j < N; ++j) {
+	for (uint32_t i = 0; i < M; ++i) {
+		for (uint32_t j = 0; j < N; ++j) {
 			lhs.elements[j][i] -= rhs.elements[j][i];
 		}
 	}
 	return lhs;
 }
 
-template<typename T, size_t M, size_t N>
+template<typename T, uint32_t M, uint32_t N>
 Matrix<T,M,N>& operator*= (Matrix<T,M,N>& lhs, T rhs) noexcept
 {
-	for (size_t i = 0; i < M; i++) {
-		for (size_t j = 0; j < N; j++) {
+	for (uint32_t i = 0; i < M; i++) {
+		for (uint32_t j = 0; j < N; j++) {
 			lhs.elements[j][i] *= rhs;
 		}
 	}
 	return lhs;
 }
 
-template<typename T, size_t N>
+template<typename T, uint32_t N>
 Matrix<T,N,N>& operator*= (Matrix<T,N,N>& lhs, const Matrix<T,N,N>& rhs) noexcept
 {
 	return (lhs = lhs * rhs);
@@ -219,36 +219,36 @@ Matrix<T,N,N>& operator*= (Matrix<T,N,N>& lhs, const Matrix<T,N,N>& rhs) noexcep
 // Operators (arithmetic)
 // ------------------------------------------------------------------------------------------------
 
-template<typename T, size_t M, size_t N>
+template<typename T, uint32_t M, uint32_t N>
 Matrix<T,M,N> operator+ (const Matrix<T,M,N>& lhs, const Matrix<T,M,N>& rhs) noexcept
 {
 	Matrix<T,M,N> temp{lhs};
 	return (temp += rhs);
 }
 
-template<typename T, size_t M, size_t N>
+template<typename T, uint32_t M, uint32_t N>
 Matrix<T,M,N> operator- (const Matrix<T,M,N>& lhs, const Matrix<T,M,N>& rhs) noexcept
 {
 	Matrix<T,M,N> temp{lhs};
 	return (temp -= rhs);
 }
 
-template<typename T, size_t M, size_t N>
+template<typename T, uint32_t M, uint32_t N>
 Matrix<T,M,N> operator- (const Matrix<T,M,N>& matrix) noexcept
 {
 	Matrix<T,M,N> temp{matrix};
 	return (temp *= T(-1));
 }
 
-template<typename T, size_t M, size_t N, size_t P>
+template<typename T, uint32_t M, uint32_t N, uint32_t P>
 Matrix<T,M,P> operator* (const Matrix<T,M,N>& lhs, const Matrix<T,N,P>& rhs) noexcept
 {
 	Matrix<T,M,P> resMatrix;
-	for (size_t i = 0; i < M; i++) {
-		for (size_t j = 0; j < P; j++) {
+	for (uint32_t i = 0; i < M; i++) {
+		for (uint32_t j = 0; j < P; j++) {
 			T temp = 0;
-			size_t jInnerThis = 0;
-			size_t iInnerOther = 0;
+			uint32_t jInnerThis = 0;
+			uint32_t iInnerOther = 0;
 			while (jInnerThis < M) {
 				temp += lhs.elements[jInnerThis][i] * rhs.elements[j][iInnerOther];
 				jInnerThis++;
@@ -260,7 +260,7 @@ Matrix<T,M,P> operator* (const Matrix<T,M,N>& lhs, const Matrix<T,N,P>& rhs) noe
 	return resMatrix;
 }
 
-template<typename T, size_t M, size_t N>
+template<typename T, uint32_t M, uint32_t N>
 Vector<T,M> operator* (const Matrix<T,M,N>& lhs, const Vector<T,N>& rhs) noexcept
 {
 	Vector<T,M> resVector;
@@ -276,14 +276,14 @@ Vector<T,M> operator* (const Matrix<T,M,N>& lhs, const Vector<T,N>& rhs) noexcep
 	return resVector;
 }
 
-template<typename T, size_t M, size_t N>
+template<typename T, uint32_t M, uint32_t N>
 Matrix<T,M,N> operator* (const Matrix<T,M,N>& lhs, T rhs) noexcept
 {
 	Matrix<T,M,N> temp{lhs};
 	return (temp *= rhs);
 }
 
-template<typename T, size_t M, size_t N>
+template<typename T, uint32_t M, uint32_t N>
 Matrix<T,M,N> operator* (T lhs, const Matrix<T,M,N>& rhs) noexcept
 {
 	return rhs * lhs;
@@ -292,11 +292,11 @@ Matrix<T,M,N> operator* (T lhs, const Matrix<T,M,N>& rhs) noexcept
 // Operators (comparison)
 // ------------------------------------------------------------------------------------------------
 
-template<typename T, size_t M, size_t N>
+template<typename T, uint32_t M, uint32_t N>
 bool operator== (const Matrix<T,M,N>& lhs, const Matrix<T,M,N>& rhs) noexcept
 {
-	for (size_t i = 0; i < M; i++) {
-		for (size_t j = 0; j < N; j++) {
+	for (uint32_t i = 0; i < M; i++) {
+		for (uint32_t j = 0; j < N; j++) {
 			if (lhs.elements[j][i] != rhs.elements[j][i]) {
 				return false;
 			}
@@ -305,7 +305,7 @@ bool operator== (const Matrix<T,M,N>& lhs, const Matrix<T,M,N>& rhs) noexcept
 	return true;
 }
 
-template<typename T, size_t M, size_t N>
+template<typename T, uint32_t M, uint32_t N>
 bool operator!= (const Matrix<T,M,N>& lhs, const Matrix<T,M,N>& rhs) noexcept
 {
 	return !(lhs == rhs);
